+++ piptools/scripts/compile.py
@@ -59,10 +59,12 @@ class PipCommand(pip.basecommand.Command):
               help="Pin packages considered unsafe: pip, setuptools & distribute")
 @click.option('--generate-hashes', is_flag=True, default=False,
               help="Generate pip 8 style hashes in the resulting requirements file.")
+@click.option('--max-rounds', default=10,
+              help="Maximum number of rounds before resolving the requirements aborts.")
 @click.argument('src_files', nargs=-1, type=click.Path(exists=True, allow_dash=True))
 def cli(verbose, dry_run, pre, rebuild, find_links, index_url, extra_index_url,
         client_cert, trusted_host, header, index, annotate, upgrade, upgrade_packages,
-        output_file, allow_unsafe, generate_hashes, src_files):
+        output_file, allow_unsafe, generate_hashes, src_files, max_rounds):
     """Compiles requirements.txt from requirements.in specs."""
     log.verbose = verbose
 
@@ -174,7 +176,7 @@ def cli(verbose, dry_run, pre, rebuild, find_links, index_url, extra_index_url,
     try:
         resolver = Resolver(constraints, repository, prereleases=pre,
                             clear_caches=rebuild)
-        results = resolver.resolve()
+        results = resolver.resolve(max_rounds=max_rounds)
         if generate_hashes:
             hashes = resolver.resolve_hashes(results)
         else: